Added static data flow analysis to track identifier dependencies.
When a TypeError occurs, we can parse the code statement at the erroneous line to determine which identifier is directly causing the TypeError. Then, we can use the identifier dependency mapping to find out the root identifier that propagates the TypeError. We assume that all TypeErrors will occur because we might have made a mistake when defining a previously undefined identifier to be of a certain type (e.g. 'int', 'NoneType' etc.)