Added default TBD class definition. Added rewrite_wrapper for function decoration. Added sample files of most common extensions. Removed data flow analysis in favor of TBD class.
(1) Added a default definition of TBD class for definitions in repsonse to NameError. The goal is to lazily define undefined variables with an object of a TBD# class. Later from the semantics of the TypeError, we will aim to unify the actual type of the lazily defined variable. (2) Added a snippet rewrite_wrapper in ActionBaseClass so that any function that is decorated with the rewrite_wrapper will intrinsically add the rewritten snippet to the code history. The requirement is that any method that is being decorated with the rewrite wrapper, needs to have access to the Snippet object. (3) We have been creating some mock files in response to FileNotFoundError. Some snippets would require the files being created to have the correct encoding of specific extensions. Therefore, samples files of most common extensions (e.g. txt, csv, xml, png and a few more) have been added to faciliatate the Create-file rule. (4) In response to TypeErrors, a dataflow analysis was previously included to match the erroneous type to the identifier having that type. This was done because the semantics of the TypeError does not reveal the identifier whose type is erroneous (e.g. 'NoneType' object is not iterable). After introducing the lazy defintion approach with TBD class, this problem is solved because the type error will then say (e.g. 'TBD0' object is not iterable). Therefore, the dataflow analysis approach to find identifier dependencies and type mappings have been removed in favor of the lazy definition approach.