Added recursive function signature discovery.
The recursive discovery recurses on Attributes, func.args and func.keywords within a function call to discover the accurate function signature. Recursing on attributes in a function call is important to discover function calls from chained function call expressions (e.g. foo.bar() or foo().bar().func() etc.). Recursing on func.args and func.keywords is important because the target function call might be an arg or a keyword arg of another function (e.g. foo(bar()) or foo.bar(func()) etc.). As a fallback for not being able to find the target function signature is to define __call__ method with a generic signature that includes variable number of positional args (i.e. *args) and variable number of keyword arguments (i.e. **kwargs). The recursive signature discovery strategy and the fallback generic signature strategy both help to resolve the unresolved TypeErrors related to . The effectiveness of these approaches can be observed from the row in Table III. that represents how many times the DefineCallable rule was applied at different iterations. Previously, the number of times DefineCallable was used at progressive iterations remained constant (i.e. 12) showing that the application of DefineCallable rule was not effectively apparent. Now, we can observe that the number of times DefineCallable rule is applied, is progressively decreasing showing that the DefineCallable rule is effective because of the recursive and the fallback signature discovery strategies.