Added two rules RemoveImport and DefineOperand. Added action seq length and venn diagram statistics in report.
The RemoveImport rule is in response to ModuleNotFoundErrors. If a module that is not found is indexed as a top PyPI package, then we attempt to install the module. Otherwise, we remove the import statement of the module and mock the module's components within the snippet as if the module's components were undefined variable, functions or attributes etc. The DefineOperand rule is in response to the TypeError related to unsupported operand type(s) for <operator>: '<class1>' and '<class2>'. In the DefineOperand rule, we try to capture the <operator> and overload the operator in either one or both of the classes '<class1>' and '<class2>'. Additionally, the report the now includes average action sequence length statistics for the total, fexec and pexec sets. The report also includes a venn diagram statistics between lexecutor and incompleter.